Multiagent HRL policy (#168)
* MultiFeedForwardPolicy -> MultiActorCriticPolicy

* added (most) code for training hierarchical multi-agent policies

* added (most) code for training hierarchical multi-agent policies

* multi_fcnet -> multiagent

* Rendered TeX expressions in 135f799de9a25b4b232d2a5c3caf923ba27eeb55

* templates for tests

* some more cleanup and bug fixes

* some bug fixes and new tests

* new tests and bug fixes

* a new way of expressing things

* bug fixes and cleanup

* bug fix

* bug fixes

* updated documentation

* Rendered TeX expressions in ce95a0843f98bdf8c4b318a0bec26fca53787e2a

* undo

* fixed a mistake

* Rendered TeX expressions in 45dc4e5391924f9a5e28887d4ec6efc2cca33e63, 636ad316be72a474984865e566bd5fd67728c7a7

* Rendered TeX expressions in 2613a3c8d7151009bfe64b8ead0fc5587b546347, 884d68cce8c2c3f4a373b733e8b19fcabab1d2a2, 7495f92dc9900ae99046a8629c0bd0af4de113c8, 1a5ea97369040a2c5bd77032e3781612931f9950

* bug fixes

* fixed merge

* added deprecated feature

* added missing file

* bug fix
