+++ experiments/run_multi_hrl.py
@@ -1,4 +1,4 @@
-"""A runner script for multi-agent fcnet models."""
+"""A runner script for multi-agent goal-conditioned hierarchical models."""
 import os
 import json
 from time import strftime
@@ -9,7 +9,7 @@ from hbaselines.utils.train import parse_options, get_hyperparameters
 from hbaselines.algorithms import OffPolicyRLAlgorithm
 
 EXAMPLE_USAGE = \
-    'python run_multi_fcnet.py "multiagent-ring-v0" --total_steps 1e6'
+    'python run_multi_hrl.py "multiagent-ring-v0" --total_steps 1e6'
 
 
 def run_exp(env,
@@ -89,20 +89,20 @@ def main(args, base_dir):
 
         # Get the policy class.
         if args.alg == "TD3":
-            from hbaselines.multiagent.td3 import MultiFeedForwardPolicy
+            from hbaselines.multiagent.h_td3 import MultiGoalConditionedPolicy
         elif args.alg == "SAC":
-            from hbaselines.multiagent.sac import MultiFeedForwardPolicy
+            from hbaselines.multiagent.h_sac import MultiGoalConditionedPolicy
         else:
             raise ValueError("Unknown algorithm: {}".format(args.alg))
 
         # Get the hyperparameters.
-        hp = get_hyperparameters(args, MultiFeedForwardPolicy)
+        hp = get_hyperparameters(args, MultiGoalConditionedPolicy)
 
         # add the seed for logging purposes
         params_with_extra = hp.copy()
         params_with_extra['seed'] = seed
         params_with_extra['env_name'] = args.env_name
-        params_with_extra['policy_name'] = "MultiFeedForwardPolicy"
+        params_with_extra['policy_name'] = "MultiGoalConditionedPolicy"
         params_with_extra['algorithm'] = args.alg
         params_with_extra['date/time'] = now
 
@@ -112,7 +112,7 @@ def main(args, base_dir):
 
         run_exp(
             env=args.env_name,
-            policy=MultiFeedForwardPolicy,
+            policy=MultiGoalConditionedPolicy,
             hp=hp,
             steps=args.total_steps,
             dir_name=dir_name,
@@ -128,10 +128,10 @@ def main(args, base_dir):
 if __name__ == '__main__':
     main(
         parse_options(
-            description='Test the performance of multi-agent fully connected '
-                        'network models on various environments.',
+            description='Test the performance of multi-agent goal-conditioned '
+                        'hierarchical models on various environments.',
             example_usage=EXAMPLE_USAGE,
             args=sys.argv[1:]
         ),
-        'data/multi-fcnet'
+        'data/multi-goal-conditioned'
     )
