+++ tests/fast_tests/test_multiagent.py
@@ -1,20 +1,26 @@
-"""Tests for the policies in the hbaselines/multi_fcnet subdirectory."""
+"""Tests for the policies in the hbaselines/multiagent subdirectory."""
 import unittest
 import numpy as np
 import tensorflow as tf
 from gym.spaces import Box
 
 from hbaselines.utils.tf_util import get_trainable_vars
-from hbaselines.multi_fcnet.td3 import MultiFeedForwardPolicy as \
+from hbaselines.multiagent.td3 import MultiFeedForwardPolicy as \
     TD3MultiFeedForwardPolicy
-from hbaselines.multi_fcnet.sac import MultiFeedForwardPolicy as \
+from hbaselines.multiagent.sac import MultiFeedForwardPolicy as \
     SACMultiFeedForwardPolicy
-from hbaselines.algorithms.off_policy import SAC_PARAMS, TD3_PARAMS
-from hbaselines.algorithms.off_policy import MULTI_FEEDFORWARD_PARAMS
+from hbaselines.multiagent.h_td3 import MultiGoalConditionedPolicy as \
+    TD3MultiGoalConditionedPolicy
+from hbaselines.multiagent.h_sac import MultiGoalConditionedPolicy as \
+    SACMultiGoalConditionedPolicy
+from hbaselines.algorithms.off_policy import SAC_PARAMS
+from hbaselines.algorithms.off_policy import TD3_PARAMS
+from hbaselines.algorithms.off_policy import MULTIAGENT_PARAMS
+from hbaselines.algorithms.off_policy import GOAL_CONDITIONED_PARAMS
 
 
 class TestMultiActorCriticPolicy(unittest.TestCase):
-    """Test MultiActorCriticPolicy in hbaselines/multi_fcnet/base.py."""
+    """Test MultiActorCriticPolicy in hbaselines/multiagent/base.py."""
 
     def setUp(self):
         self.sess = tf.compat.v1.Session()
@@ -30,7 +36,7 @@ class TestMultiActorCriticPolicy(unittest.TestCase):
             'verbose': 0,
         }
         self.policy_params_shared.update(TD3_PARAMS.copy())
-        self.policy_params_shared.update(MULTI_FEEDFORWARD_PARAMS.copy())
+        self.policy_params_shared.update(MULTIAGENT_PARAMS.copy())
         self.policy_params_shared['shared'] = True
 
         # Independent policy parameters
@@ -53,7 +59,7 @@ class TestMultiActorCriticPolicy(unittest.TestCase):
             'verbose': 0,
         }
         self.policy_params_independent.update(TD3_PARAMS.copy())
-        self.policy_params_independent.update(MULTI_FEEDFORWARD_PARAMS.copy())
+        self.policy_params_independent.update(MULTIAGENT_PARAMS.copy())
         self.policy_params_independent['shared'] = False
 
     def tearDown(self):
@@ -247,7 +253,7 @@ class TestMultiActorCriticPolicy(unittest.TestCase):
 
 
 class TestTD3MultiFeedForwardPolicy(unittest.TestCase):
-    """Test MultiFeedForwardPolicy in hbaselines/multi_fcnet/td3.py."""
+    """Test MultiFeedForwardPolicy in hbaselines/multiagent/td3.py."""
 
     def setUp(self):
         self.sess = tf.compat.v1.Session()
@@ -263,7 +269,7 @@ class TestTD3MultiFeedForwardPolicy(unittest.TestCase):
             'verbose': 0,
         }
         self.policy_params_shared.update(TD3_PARAMS.copy())
-        self.policy_params_shared.update(MULTI_FEEDFORWARD_PARAMS.copy())
+        self.policy_params_shared.update(MULTIAGENT_PARAMS.copy())
         self.policy_params_shared['shared'] = True
 
         # Independent policy parameters
@@ -286,7 +292,7 @@ class TestTD3MultiFeedForwardPolicy(unittest.TestCase):
             'verbose': 0,
         }
         self.policy_params_independent.update(TD3_PARAMS.copy())
-        self.policy_params_independent.update(MULTI_FEEDFORWARD_PARAMS.copy())
+        self.policy_params_independent.update(MULTIAGENT_PARAMS.copy())
         self.policy_params_independent['shared'] = False
 
     def tearDown(self):
@@ -296,6 +302,18 @@ class TestTD3MultiFeedForwardPolicy(unittest.TestCase):
 
         # Clear the graph.
         tf.compat.v1.reset_default_graph()
+
+    def test_deprecated(self):
+        """Make sure that the original path still works (temporarily)."""
+        raised = False
+        try:
+            from hbaselines.multi_fcnet.td3 import MultiFeedForwardPolicy
+            policy_params = self.policy_params_independent.copy()
+            _ = MultiFeedForwardPolicy(**policy_params)
+        except ModuleNotFoundError:  # pragma: no cover
+            raised = True  # pragma: no cover
+
+        self.assertFalse(raised, 'Exception raised')
 
     def test_init_1(self):
         """Check the functionality of the __init__() method.
@@ -1183,7 +1201,7 @@ class TestTD3MultiFeedForwardPolicy(unittest.TestCase):
 
 
 class TestSACMultiFeedForwardPolicy(unittest.TestCase):
-    """Test MultiFeedForwardPolicy in hbaselines/multi_fcnet/sac.py."""
+    """Test MultiFeedForwardPolicy in hbaselines/multiagent/sac.py."""
 
     def setUp(self):
         self.sess = tf.compat.v1.Session()
@@ -1199,7 +1217,7 @@ class TestSACMultiFeedForwardPolicy(unittest.TestCase):
             'verbose': 0,
         }
         self.policy_params_shared.update(SAC_PARAMS.copy())
-        self.policy_params_shared.update(MULTI_FEEDFORWARD_PARAMS.copy())
+        self.policy_params_shared.update(MULTIAGENT_PARAMS.copy())
         self.policy_params_shared['shared'] = True
 
         # Independent policy parameters
@@ -1222,7 +1240,7 @@ class TestSACMultiFeedForwardPolicy(unittest.TestCase):
             'verbose': 0,
         }
         self.policy_params_independent.update(SAC_PARAMS.copy())
-        self.policy_params_independent.update(MULTI_FEEDFORWARD_PARAMS.copy())
+        self.policy_params_independent.update(MULTIAGENT_PARAMS.copy())
         self.policy_params_independent['shared'] = False
 
     def tearDown(self):
@@ -1232,6 +1250,18 @@ class TestSACMultiFeedForwardPolicy(unittest.TestCase):
 
         # Clear the graph.
         tf.compat.v1.reset_default_graph()
+
+    def test_deprecated(self):
+        """Make sure that the original path still works (temporarily)."""
+        raised = False
+        try:
+            from hbaselines.multi_fcnet.sac import MultiFeedForwardPolicy
+            policy_params = self.policy_params_independent.copy()
+            _ = MultiFeedForwardPolicy(**policy_params)
+        except ModuleNotFoundError:  # pragma: no cover
+            raised = True  # pragma: no cover
+
+        self.assertFalse(raised, 'Exception raised')
 
     def test_init_1(self):
         """Check the functionality of the __init__() method.
@@ -1725,6 +1755,1405 @@ class TestSACMultiFeedForwardPolicy(unittest.TestCase):
             'target/centralized_value_fns/vf/fc1/kernel:0',
             'target/centralized_value_fns/vf/vf_output/bias:0',
             'target/centralized_value_fns/vf/vf_output/kernel:0',
+        ]
+
+        for model, target in zip(model_var_list, target_var_list):
+            with tf.compat.v1.variable_scope(
+                    tf.compat.v1.get_variable_scope(), reuse=True):
+                model_val = policy.sess.run(model)
+                target_val = policy.sess.run(target)
+            np.testing.assert_almost_equal(model_val, target_val)
+
+    def test_store_transition_1(self):
+        """Check the functionality of the store_transition() method.
+
+        This test checks for the following cases:
+
+        1. maddpg = True,  shared = False
+        2. maddpg = True,  shared = True
+        """
+        policy_params = self.policy_params_independent.copy()
+        policy_params["maddpg"] = True
+        policy = SACMultiFeedForwardPolicy(**policy_params)
+
+        # Initialize the variables of the policy.
+        policy.sess.run(tf.compat.v1.global_variables_initializer())
+
+        # Run the initialize method.
+        policy.initialize()
+
+        for i in range(4):
+            action_0 = np.array([i for _ in range(1)])
+            action_1 = np.array([i for _ in range(2)])
+            context0_0 = np.array([i for _ in range(3)])
+            context0_1 = np.array([i for _ in range(4)])
+            obs0_0 = np.array([i for _ in range(5)])
+            obs0_1 = np.array([i for _ in range(6)])
+            reward = i
+            obs1_0 = np.array([i+1 for _ in range(5)])
+            obs1_1 = np.array([i+1 for _ in range(6)])
+            context1_0 = np.array([i for _ in range(3)])
+            context1_1 = np.array([i for _ in range(4)])
+            done = False
+            is_final_step = False
+            evaluate = False
+            all_obs0 = np.array([i for _ in range(18)])
+            all_obs1 = np.array([i+1 for _ in range(18)])
+
+            policy.store_transition(
+                obs0={"a": obs0_0, "b": obs0_1},
+                context0={"a": context0_0, "b": context0_1},
+                action={"a": action_0, "b": action_1},
+                reward={"a": reward, "b": reward},
+                obs1={"a": obs1_0, "b": obs1_1},
+                context1={"a": context1_0, "b": context1_1},
+                done=done,
+                is_final_step=is_final_step,
+                evaluate=evaluate,
+                env_num=0,
+                all_obs0=all_obs0,
+                all_obs1=all_obs1,
+            )
+
+        # =================================================================== #
+        # test for agent a                                                    #
+        # =================================================================== #
+
+        obs_t = policy.replay_buffer["a"].obs_t
+        action_t = policy.replay_buffer["a"].action_t
+        reward = policy.replay_buffer["a"].reward
+        done = policy.replay_buffer["a"].done
+        all_obs_t = policy.replay_buffer["a"].all_obs_t
+
+        # check the various attributes
+        np.testing.assert_almost_equal(
+            obs_t[:4, :],
+            np.array([[0., 0., 0., 0., 0., 0., 0., 0.],
+                      [1., 1., 1., 1., 1., 1., 1., 1.],
+                      [2., 2., 2., 2., 2., 2., 2., 2.],
+                      [3., 3., 3., 3., 3., 3., 3., 3.]])
+        )
+
+        np.testing.assert_almost_equal(
+            action_t[:4, :],
+            np.array([[0.], [1.], [2.], [3.]])
+        )
+
+        np.testing.assert_almost_equal(
+            reward[:4],
+            np.array([0., 1., 2., 3.])
+        )
+
+        np.testing.assert_almost_equal(
+            done[:4],
+            [0., 0., 0., 0.]
+        )
+
+        np.testing.assert_almost_equal(
+            all_obs_t[:4, :],
+            np.array([[0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0.,
+                       0., 0., 0., 0.],
+                      [1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.,
+                       1., 1., 1., 1.],
+                      [2., 2., 2., 2., 2., 2., 2., 2., 2., 2., 2., 2., 2., 2.,
+                       2., 2., 2., 2.],
+                      [3., 3., 3., 3., 3., 3., 3., 3., 3., 3., 3., 3., 3., 3.,
+                       3., 3., 3., 3.]])
+        )
+
+        # =================================================================== #
+        # test for agent b                                                    #
+        # =================================================================== #
+
+        obs_t = policy.replay_buffer["b"].obs_t
+        action_t = policy.replay_buffer["b"].action_t
+        reward = policy.replay_buffer["b"].reward
+        done = policy.replay_buffer["b"].done
+        all_obs_t = policy.replay_buffer["b"].all_obs_t
+
+        # check the various attributes
+        np.testing.assert_almost_equal(
+            obs_t[:4, :],
+            np.array([[0., 0., 0., 0., 0., 0., 0., 0., 0., 0.],
+                      [1., 1., 1., 1., 1., 1., 1., 1., 1., 1.],
+                      [2., 2., 2., 2., 2., 2., 2., 2., 2., 2.],
+                      [3., 3., 3., 3., 3., 3., 3., 3., 3., 3.]])
+        )
+
+        np.testing.assert_almost_equal(
+            action_t[:4, :],
+            np.array([[0., 0.], [1., 1.], [2., 2.], [3., 3.]])
+        )
+
+        np.testing.assert_almost_equal(
+            reward[:4],
+            np.array([0., 1., 2., 3.])
+        )
+
+        np.testing.assert_almost_equal(
+            done[:4],
+            [0., 0., 0., 0.]
+        )
+
+        np.testing.assert_almost_equal(
+            all_obs_t[:4, :],
+            np.array([[0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0.,
+                       0., 0., 0., 0.],
+                      [1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.,
+                       1., 1., 1., 1.],
+                      [2., 2., 2., 2., 2., 2., 2., 2., 2., 2., 2., 2., 2., 2.,
+                       2., 2., 2., 2.],
+                      [3., 3., 3., 3., 3., 3., 3., 3., 3., 3., 3., 3., 3., 3.,
+                       3., 3., 3., 3.]])
+        )
+
+    def test_store_transition_4(self):
+        policy_params = self.policy_params_shared.copy()
+        policy_params["maddpg"] = True
+        policy_params["n_agents"] = 2
+        policy = SACMultiFeedForwardPolicy(**policy_params)
+
+        # Initialize the variables of the policy.
+        policy.sess.run(tf.compat.v1.global_variables_initializer())
+
+        # Run the initialize method.
+        policy.initialize()
+
+        for i in range(4):
+            obs0 = np.array([i for _ in range(2)])
+            context0 = np.array([i for _ in range(3)])
+            action = np.array([i for _ in range(1)])
+            reward = i
+            obs1 = np.array([i+1 for _ in range(2)])
+            context1 = np.array([i for _ in range(3)])
+            is_final_step = False
+            evaluate = False
+            all_obs0 = np.array([i for _ in range(10)])
+            all_obs1 = np.array([i+1 for _ in range(10)])
+
+            policy.store_transition(
+                obs0={"a": obs0, "b": obs0 + 1},
+                context0={"a": context0, "b": context0 + 1},
+                action={"a": action, "b": action + 1},
+                reward={"a": reward, "b": reward + 1},
+                obs1={"a": obs1, "b": obs1 + 1},
+                context1={"a": context1, "b": context1 + 1},
+                done=0.,
+                is_final_step=is_final_step,
+                evaluate=evaluate,
+                env_num=0,
+                all_obs0=all_obs0,
+                all_obs1=all_obs1,
+            )
+
+        # extract the attributes
+        obs_t = policy.replay_buffer.obs_t
+        action_t = policy.replay_buffer.action
+        reward = policy.replay_buffer.reward
+        done = policy.replay_buffer.done
+        all_obs_t = policy.replay_buffer.all_obs_t
+
+        # check the various attributes
+        np.testing.assert_almost_equal(
+            obs_t[0][:4, :],
+            np.array([[0., 0., 0., 0., 0.],
+                      [1., 1., 1., 1., 1.],
+                      [2., 2., 2., 2., 2.],
+                      [3., 3., 3., 3., 3.]])
+        )
+
+        np.testing.assert_almost_equal(
+            action_t[0][:4, :],
+            np.array([[0.], [1.], [2.], [3.]])
+        )
+
+        np.testing.assert_almost_equal(
+            reward[:4],
+            np.array([0., 1., 2., 3.])
+        )
+
+        np.testing.assert_almost_equal(
+            done[:4],
+            [0., 0., 0., 0.]
+        )
+
+        np.testing.assert_almost_equal(
+            all_obs_t[:4, :],
+            np.array([[0., 0., 0., 0., 0., 0., 0., 0., 0., 0.],
+                      [1., 1., 1., 1., 1., 1., 1., 1., 1., 1.],
+                      [2., 2., 2., 2., 2., 2., 2., 2., 2., 2.],
+                      [3., 3., 3., 3., 3., 3., 3., 3., 3., 3.]])
+        )
+
+
+class TestTD3MultiGoalConditionedPolicy(unittest.TestCase):
+    """Test MultiFeedForwardPolicy in hbaselines/multiagent/h_td3.py."""
+
+    def setUp(self):
+        self.sess = tf.compat.v1.Session()
+
+        # Shared policy parameters
+        self.policy_params_shared = {
+            'sess': self.sess,
+            'ac_space': Box(low=-1, high=1, shape=(1,)),
+            'co_space': Box(low=-2, high=2, shape=(2,)),
+            'ob_space': Box(low=-3, high=3, shape=(3,)),
+            'all_ob_space': Box(low=-3, high=3, shape=(10,)),
+            'layers': [256, 256],
+            'verbose': 0,
+        }
+        self.policy_params_shared.update(TD3_PARAMS.copy())
+        self.policy_params_shared.update(GOAL_CONDITIONED_PARAMS.copy())
+        self.policy_params_shared.update(MULTIAGENT_PARAMS.copy())
+        self.policy_params_shared['shared'] = True
+
+        # Independent policy parameters
+        self.policy_params_independent = {
+            'sess': self.sess,
+            'ac_space': {
+                'a': Box(low=-1, high=1, shape=(1,)),
+                'b': Box(low=-2, high=2, shape=(2,)),
+            },
+            'co_space': {
+                'a': Box(low=-3, high=3, shape=(3,)),
+                'b': Box(low=-4, high=4, shape=(4,)),
+            },
+            'ob_space': {
+                'a': Box(low=-5, high=5, shape=(5,)),
+                'b': Box(low=-6, high=6, shape=(6,)),
+            },
+            'all_ob_space': Box(low=-6, high=6, shape=(18,)),
+            'layers': [256, 256],
+            'verbose': 0,
+        }
+        self.policy_params_independent.update(TD3_PARAMS.copy())
+        self.policy_params_independent.update(GOAL_CONDITIONED_PARAMS.copy())
+        self.policy_params_independent.update(MULTIAGENT_PARAMS.copy())
+        self.policy_params_independent['shared'] = False
+
+    def tearDown(self):
+        self.sess.close()
+        del self.policy_params_shared
+        del self.policy_params_independent
+
+        # Clear the graph.
+        tf.compat.v1.reset_default_graph()
+
+    def test_init_1(self):
+        """Check the functionality of the __init__() method.
+
+        This method is tested for the following features:
+
+        1. The proper structure graph was generated.
+        2. All input placeholders are correct.
+
+        This is done for the following cases:
+
+        1. maddpg = False, shared = False
+        2. maddpg = False, shared = True
+        """
+        policy_params = self.policy_params_independent.copy()
+        policy_params["maddpg"] = False
+        policy = TD3MultiGoalConditionedPolicy(**policy_params)
+
+        self.assertListEqual(
+            sorted([var.name for var in get_trainable_vars()]),
+            ['a/level_0/model/pi/fc0/bias:0',
+             'a/level_0/model/pi/fc0/kernel:0',
+             'a/level_0/model/pi/fc1/bias:0',
+             'a/level_0/model/pi/fc1/kernel:0',
+             'a/level_0/model/pi/output/bias:0',
+             'a/level_0/model/pi/output/kernel:0',
+             'a/level_0/model/qf_0/fc0/bias:0',
+             'a/level_0/model/qf_0/fc0/kernel:0',
+             'a/level_0/model/qf_0/fc1/bias:0',
+             'a/level_0/model/qf_0/fc1/kernel:0',
+             'a/level_0/model/qf_0/qf_output/bias:0',
+             'a/level_0/model/qf_0/qf_output/kernel:0',
+             'a/level_0/model/qf_1/fc0/bias:0',
+             'a/level_0/model/qf_1/fc0/kernel:0',
+             'a/level_0/model/qf_1/fc1/bias:0',
+             'a/level_0/model/qf_1/fc1/kernel:0',
+             'a/level_0/model/qf_1/qf_output/bias:0',
+             'a/level_0/model/qf_1/qf_output/kernel:0',
+             'a/level_0/target/pi/fc0/bias:0',
+             'a/level_0/target/pi/fc0/kernel:0',
+             'a/level_0/target/pi/fc1/bias:0',
+             'a/level_0/target/pi/fc1/kernel:0',
+             'a/level_0/target/pi/output/bias:0',
+             'a/level_0/target/pi/output/kernel:0',
+             'a/level_0/target/qf_0/fc0/bias:0',
+             'a/level_0/target/qf_0/fc0/kernel:0',
+             'a/level_0/target/qf_0/fc1/bias:0',
+             'a/level_0/target/qf_0/fc1/kernel:0',
+             'a/level_0/target/qf_0/qf_output/bias:0',
+             'a/level_0/target/qf_0/qf_output/kernel:0',
+             'a/level_0/target/qf_1/fc0/bias:0',
+             'a/level_0/target/qf_1/fc0/kernel:0',
+             'a/level_0/target/qf_1/fc1/bias:0',
+             'a/level_0/target/qf_1/fc1/kernel:0',
+             'a/level_0/target/qf_1/qf_output/bias:0',
+             'a/level_0/target/qf_1/qf_output/kernel:0',
+             'a/level_1/model/pi/fc0/bias:0',
+             'a/level_1/model/pi/fc0/kernel:0',
+             'a/level_1/model/pi/fc1/bias:0',
+             'a/level_1/model/pi/fc1/kernel:0',
+             'a/level_1/model/pi/output/bias:0',
+             'a/level_1/model/pi/output/kernel:0',
+             'a/level_1/model/qf_0/fc0/bias:0',
+             'a/level_1/model/qf_0/fc0/kernel:0',
+             'a/level_1/model/qf_0/fc1/bias:0',
+             'a/level_1/model/qf_0/fc1/kernel:0',
+             'a/level_1/model/qf_0/qf_output/bias:0',
+             'a/level_1/model/qf_0/qf_output/kernel:0',
+             'a/level_1/model/qf_1/fc0/bias:0',
+             'a/level_1/model/qf_1/fc0/kernel:0',
+             'a/level_1/model/qf_1/fc1/bias:0',
+             'a/level_1/model/qf_1/fc1/kernel:0',
+             'a/level_1/model/qf_1/qf_output/bias:0',
+             'a/level_1/model/qf_1/qf_output/kernel:0',
+             'a/level_1/target/pi/fc0/bias:0',
+             'a/level_1/target/pi/fc0/kernel:0',
+             'a/level_1/target/pi/fc1/bias:0',
+             'a/level_1/target/pi/fc1/kernel:0',
+             'a/level_1/target/pi/output/bias:0',
+             'a/level_1/target/pi/output/kernel:0',
+             'a/level_1/target/qf_0/fc0/bias:0',
+             'a/level_1/target/qf_0/fc0/kernel:0',
+             'a/level_1/target/qf_0/fc1/bias:0',
+             'a/level_1/target/qf_0/fc1/kernel:0',
+             'a/level_1/target/qf_0/qf_output/bias:0',
+             'a/level_1/target/qf_0/qf_output/kernel:0',
+             'a/level_1/target/qf_1/fc0/bias:0',
+             'a/level_1/target/qf_1/fc0/kernel:0',
+             'a/level_1/target/qf_1/fc1/bias:0',
+             'a/level_1/target/qf_1/fc1/kernel:0',
+             'a/level_1/target/qf_1/qf_output/bias:0',
+             'a/level_1/target/qf_1/qf_output/kernel:0',
+             'b/level_0/model/pi/fc0/bias:0',
+             'b/level_0/model/pi/fc0/kernel:0',
+             'b/level_0/model/pi/fc1/bias:0',
+             'b/level_0/model/pi/fc1/kernel:0',
+             'b/level_0/model/pi/output/bias:0',
+             'b/level_0/model/pi/output/kernel:0',
+             'b/level_0/model/qf_0/fc0/bias:0',
+             'b/level_0/model/qf_0/fc0/kernel:0',
+             'b/level_0/model/qf_0/fc1/bias:0',
+             'b/level_0/model/qf_0/fc1/kernel:0',
+             'b/level_0/model/qf_0/qf_output/bias:0',
+             'b/level_0/model/qf_0/qf_output/kernel:0',
+             'b/level_0/model/qf_1/fc0/bias:0',
+             'b/level_0/model/qf_1/fc0/kernel:0',
+             'b/level_0/model/qf_1/fc1/bias:0',
+             'b/level_0/model/qf_1/fc1/kernel:0',
+             'b/level_0/model/qf_1/qf_output/bias:0',
+             'b/level_0/model/qf_1/qf_output/kernel:0',
+             'b/level_0/target/pi/fc0/bias:0',
+             'b/level_0/target/pi/fc0/kernel:0',
+             'b/level_0/target/pi/fc1/bias:0',
+             'b/level_0/target/pi/fc1/kernel:0',
+             'b/level_0/target/pi/output/bias:0',
+             'b/level_0/target/pi/output/kernel:0',
+             'b/level_0/target/qf_0/fc0/bias:0',
+             'b/level_0/target/qf_0/fc0/kernel:0',
+             'b/level_0/target/qf_0/fc1/bias:0',
+             'b/level_0/target/qf_0/fc1/kernel:0',
+             'b/level_0/target/qf_0/qf_output/bias:0',
+             'b/level_0/target/qf_0/qf_output/kernel:0',
+             'b/level_0/target/qf_1/fc0/bias:0',
+             'b/level_0/target/qf_1/fc0/kernel:0',
+             'b/level_0/target/qf_1/fc1/bias:0',
+             'b/level_0/target/qf_1/fc1/kernel:0',
+             'b/level_0/target/qf_1/qf_output/bias:0',
+             'b/level_0/target/qf_1/qf_output/kernel:0',
+             'b/level_1/model/pi/fc0/bias:0',
+             'b/level_1/model/pi/fc0/kernel:0',
+             'b/level_1/model/pi/fc1/bias:0',
+             'b/level_1/model/pi/fc1/kernel:0',
+             'b/level_1/model/pi/output/bias:0',
+             'b/level_1/model/pi/output/kernel:0',
+             'b/level_1/model/qf_0/fc0/bias:0',
+             'b/level_1/model/qf_0/fc0/kernel:0',
+             'b/level_1/model/qf_0/fc1/bias:0',
+             'b/level_1/model/qf_0/fc1/kernel:0',
+             'b/level_1/model/qf_0/qf_output/bias:0',
+             'b/level_1/model/qf_0/qf_output/kernel:0',
+             'b/level_1/model/qf_1/fc0/bias:0',
+             'b/level_1/model/qf_1/fc0/kernel:0',
+             'b/level_1/model/qf_1/fc1/bias:0',
+             'b/level_1/model/qf_1/fc1/kernel:0',
+             'b/level_1/model/qf_1/qf_output/bias:0',
+             'b/level_1/model/qf_1/qf_output/kernel:0',
+             'b/level_1/target/pi/fc0/bias:0',
+             'b/level_1/target/pi/fc0/kernel:0',
+             'b/level_1/target/pi/fc1/bias:0',
+             'b/level_1/target/pi/fc1/kernel:0',
+             'b/level_1/target/pi/output/bias:0',
+             'b/level_1/target/pi/output/kernel:0',
+             'b/level_1/target/qf_0/fc0/bias:0',
+             'b/level_1/target/qf_0/fc0/kernel:0',
+             'b/level_1/target/qf_0/fc1/bias:0',
+             'b/level_1/target/qf_0/fc1/kernel:0',
+             'b/level_1/target/qf_0/qf_output/bias:0',
+             'b/level_1/target/qf_0/qf_output/kernel:0',
+             'b/level_1/target/qf_1/fc0/bias:0',
+             'b/level_1/target/qf_1/fc0/kernel:0',
+             'b/level_1/target/qf_1/fc1/bias:0',
+             'b/level_1/target/qf_1/fc1/kernel:0',
+             'b/level_1/target/qf_1/qf_output/bias:0',
+             'b/level_1/target/qf_1/qf_output/kernel:0']
+        )
+
+        # Check observation/action/context spaces of the agents
+        self.assertEqual(
+            policy.agents['a'].policy[0].ac_space.shape[0],
+            self.policy_params_independent['ob_space']['a'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['a'].policy[0].ob_space.shape[0],
+            self.policy_params_independent['ob_space']['a'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['a'].policy[0].co_space.shape[0],
+            self.policy_params_independent['co_space']['a'].shape[0]
+        )
+
+        self.assertEqual(
+            policy.agents['a'].policy[1].ac_space.shape[0],
+            self.policy_params_independent['ac_space']['a'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['a'].policy[1].ob_space.shape[0],
+            self.policy_params_independent['ob_space']['a'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['a'].policy[1].co_space.shape[0],
+            self.policy_params_independent['ob_space']['a'].shape[0]
+        )
+
+        self.assertEqual(
+            policy.agents['b'].policy[0].ac_space.shape[0],
+            self.policy_params_independent['ob_space']['b'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['b'].policy[0].ob_space.shape[0],
+            self.policy_params_independent['ob_space']['b'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['b'].policy[0].co_space.shape[0],
+            self.policy_params_independent['co_space']['b'].shape[0]
+        )
+
+        self.assertEqual(
+            policy.agents['b'].policy[1].ac_space.shape[0],
+            self.policy_params_independent['ac_space']['b'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['b'].policy[1].ob_space.shape[0],
+            self.policy_params_independent['ob_space']['b'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['b'].policy[1].co_space.shape[0],
+            self.policy_params_independent['ob_space']['b'].shape[0]
+        )
+
+        # Check the instantiation of the class attributes.
+        self.assertTrue(not policy.shared)
+        self.assertTrue(not policy.maddpg)
+
+    def test_init_2(self):
+        policy_params = self.policy_params_shared.copy()
+        policy_params["maddpg"] = False
+        policy = TD3MultiGoalConditionedPolicy(**policy_params)
+
+        self.assertListEqual(
+            sorted([var.name for var in get_trainable_vars()]),
+            ['level_0/model/pi/fc0/bias:0',
+             'level_0/model/pi/fc0/kernel:0',
+             'level_0/model/pi/fc1/bias:0',
+             'level_0/model/pi/fc1/kernel:0',
+             'level_0/model/pi/output/bias:0',
+             'level_0/model/pi/output/kernel:0',
+             'level_0/model/qf_0/fc0/bias:0',
+             'level_0/model/qf_0/fc0/kernel:0',
+             'level_0/model/qf_0/fc1/bias:0',
+             'level_0/model/qf_0/fc1/kernel:0',
+             'level_0/model/qf_0/qf_output/bias:0',
+             'level_0/model/qf_0/qf_output/kernel:0',
+             'level_0/model/qf_1/fc0/bias:0',
+             'level_0/model/qf_1/fc0/kernel:0',
+             'level_0/model/qf_1/fc1/bias:0',
+             'level_0/model/qf_1/fc1/kernel:0',
+             'level_0/model/qf_1/qf_output/bias:0',
+             'level_0/model/qf_1/qf_output/kernel:0',
+             'level_0/target/pi/fc0/bias:0',
+             'level_0/target/pi/fc0/kernel:0',
+             'level_0/target/pi/fc1/bias:0',
+             'level_0/target/pi/fc1/kernel:0',
+             'level_0/target/pi/output/bias:0',
+             'level_0/target/pi/output/kernel:0',
+             'level_0/target/qf_0/fc0/bias:0',
+             'level_0/target/qf_0/fc0/kernel:0',
+             'level_0/target/qf_0/fc1/bias:0',
+             'level_0/target/qf_0/fc1/kernel:0',
+             'level_0/target/qf_0/qf_output/bias:0',
+             'level_0/target/qf_0/qf_output/kernel:0',
+             'level_0/target/qf_1/fc0/bias:0',
+             'level_0/target/qf_1/fc0/kernel:0',
+             'level_0/target/qf_1/fc1/bias:0',
+             'level_0/target/qf_1/fc1/kernel:0',
+             'level_0/target/qf_1/qf_output/bias:0',
+             'level_0/target/qf_1/qf_output/kernel:0',
+             'level_1/model/pi/fc0/bias:0',
+             'level_1/model/pi/fc0/kernel:0',
+             'level_1/model/pi/fc1/bias:0',
+             'level_1/model/pi/fc1/kernel:0',
+             'level_1/model/pi/output/bias:0',
+             'level_1/model/pi/output/kernel:0',
+             'level_1/model/qf_0/fc0/bias:0',
+             'level_1/model/qf_0/fc0/kernel:0',
+             'level_1/model/qf_0/fc1/bias:0',
+             'level_1/model/qf_0/fc1/kernel:0',
+             'level_1/model/qf_0/qf_output/bias:0',
+             'level_1/model/qf_0/qf_output/kernel:0',
+             'level_1/model/qf_1/fc0/bias:0',
+             'level_1/model/qf_1/fc0/kernel:0',
+             'level_1/model/qf_1/fc1/bias:0',
+             'level_1/model/qf_1/fc1/kernel:0',
+             'level_1/model/qf_1/qf_output/bias:0',
+             'level_1/model/qf_1/qf_output/kernel:0',
+             'level_1/target/pi/fc0/bias:0',
+             'level_1/target/pi/fc0/kernel:0',
+             'level_1/target/pi/fc1/bias:0',
+             'level_1/target/pi/fc1/kernel:0',
+             'level_1/target/pi/output/bias:0',
+             'level_1/target/pi/output/kernel:0',
+             'level_1/target/qf_0/fc0/bias:0',
+             'level_1/target/qf_0/fc0/kernel:0',
+             'level_1/target/qf_0/fc1/bias:0',
+             'level_1/target/qf_0/fc1/kernel:0',
+             'level_1/target/qf_0/qf_output/bias:0',
+             'level_1/target/qf_0/qf_output/kernel:0',
+             'level_1/target/qf_1/fc0/bias:0',
+             'level_1/target/qf_1/fc0/kernel:0',
+             'level_1/target/qf_1/fc1/bias:0',
+             'level_1/target/qf_1/fc1/kernel:0',
+             'level_1/target/qf_1/qf_output/bias:0',
+             'level_1/target/qf_1/qf_output/kernel:0']
+        )
+
+        # Check observation/action/context spaces of the agents
+        self.assertEqual(
+            policy.agents['policy'].policy[0].ac_space.shape[0],
+            self.policy_params_shared['ob_space'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['policy'].policy[0].ob_space.shape[0],
+            self.policy_params_shared['ob_space'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['policy'].policy[0].co_space.shape[0],
+            self.policy_params_shared['co_space'].shape[0]
+        )
+
+        self.assertEqual(
+            policy.agents['policy'].policy[1].ac_space.shape[0],
+            self.policy_params_shared['ac_space'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['policy'].policy[1].ob_space.shape[0],
+            self.policy_params_shared['ob_space'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['policy'].policy[1].co_space.shape[0],
+            self.policy_params_shared['ob_space'].shape[0]
+        )
+
+        # Check the instantiation of the class attributes.
+        self.assertTrue(policy.shared)
+        self.assertTrue(not policy.maddpg)
+
+    def test_initialize_1(self):
+        """Check the functionality of the initialize() method.
+
+        This test validates that the target variables are properly initialized
+        when initialize is called.
+
+        This is done for the following cases:
+
+        1. maddpg = False, shared = False
+        2. maddpg = False, shared = True
+        """
+        policy_params = self.policy_params_independent.copy()
+        policy_params["maddpg"] = False
+        policy = TD3MultiGoalConditionedPolicy(**policy_params)
+
+        # Initialize the variables of the policy.
+        policy.sess.run(tf.compat.v1.global_variables_initializer())
+
+        # Run the initialize method.
+        policy.initialize()
+
+        model_var_list = [
+            'a/level_0/model/pi/fc0/bias:0',
+            'a/level_0/model/pi/fc0/kernel:0',
+            'a/level_0/model/pi/fc1/bias:0',
+            'a/level_0/model/pi/fc1/kernel:0',
+            'a/level_0/model/pi/output/bias:0',
+            'a/level_0/model/pi/output/kernel:0',
+            'a/level_0/model/qf_0/fc0/bias:0',
+            'a/level_0/model/qf_0/fc0/kernel:0',
+            'a/level_0/model/qf_0/fc1/bias:0',
+            'a/level_0/model/qf_0/fc1/kernel:0',
+            'a/level_0/model/qf_0/qf_output/bias:0',
+            'a/level_0/model/qf_0/qf_output/kernel:0',
+            'a/level_0/model/qf_1/fc0/bias:0',
+            'a/level_0/model/qf_1/fc0/kernel:0',
+            'a/level_0/model/qf_1/fc1/bias:0',
+            'a/level_0/model/qf_1/fc1/kernel:0',
+            'a/level_0/model/qf_1/qf_output/bias:0',
+            'a/level_0/model/qf_1/qf_output/kernel:0',
+
+            'a/level_1/model/pi/fc0/bias:0',
+            'a/level_1/model/pi/fc0/kernel:0',
+            'a/level_1/model/pi/fc1/bias:0',
+            'a/level_1/model/pi/fc1/kernel:0',
+            'a/level_1/model/pi/output/bias:0',
+            'a/level_1/model/pi/output/kernel:0',
+            'a/level_1/model/qf_0/fc0/bias:0',
+            'a/level_1/model/qf_0/fc0/kernel:0',
+            'a/level_1/model/qf_0/fc1/bias:0',
+            'a/level_1/model/qf_0/fc1/kernel:0',
+            'a/level_1/model/qf_0/qf_output/bias:0',
+            'a/level_1/model/qf_0/qf_output/kernel:0',
+            'a/level_1/model/qf_1/fc0/bias:0',
+            'a/level_1/model/qf_1/fc0/kernel:0',
+            'a/level_1/model/qf_1/fc1/bias:0',
+            'a/level_1/model/qf_1/fc1/kernel:0',
+            'a/level_1/model/qf_1/qf_output/bias:0',
+            'a/level_1/model/qf_1/qf_output/kernel:0',
+
+            'b/level_0/model/pi/fc0/bias:0',
+            'b/level_0/model/pi/fc0/kernel:0',
+            'b/level_0/model/pi/fc1/bias:0',
+            'b/level_0/model/pi/fc1/kernel:0',
+            'b/level_0/model/pi/output/bias:0',
+            'b/level_0/model/pi/output/kernel:0',
+            'b/level_0/model/qf_0/fc0/bias:0',
+            'b/level_0/model/qf_0/fc0/kernel:0',
+            'b/level_0/model/qf_0/fc1/bias:0',
+            'b/level_0/model/qf_0/fc1/kernel:0',
+            'b/level_0/model/qf_0/qf_output/bias:0',
+            'b/level_0/model/qf_0/qf_output/kernel:0',
+            'b/level_0/model/qf_1/fc0/bias:0',
+            'b/level_0/model/qf_1/fc0/kernel:0',
+            'b/level_0/model/qf_1/fc1/bias:0',
+            'b/level_0/model/qf_1/fc1/kernel:0',
+            'b/level_0/model/qf_1/qf_output/bias:0',
+            'b/level_0/model/qf_1/qf_output/kernel:0',
+
+            'b/level_1/model/pi/fc0/bias:0',
+            'b/level_1/model/pi/fc0/kernel:0',
+            'b/level_1/model/pi/fc1/bias:0',
+            'b/level_1/model/pi/fc1/kernel:0',
+            'b/level_1/model/pi/output/bias:0',
+            'b/level_1/model/pi/output/kernel:0',
+            'b/level_1/model/qf_0/fc0/bias:0',
+            'b/level_1/model/qf_0/fc0/kernel:0',
+            'b/level_1/model/qf_0/fc1/bias:0',
+            'b/level_1/model/qf_0/fc1/kernel:0',
+            'b/level_1/model/qf_0/qf_output/bias:0',
+            'b/level_1/model/qf_0/qf_output/kernel:0',
+            'b/level_1/model/qf_1/fc0/bias:0',
+            'b/level_1/model/qf_1/fc0/kernel:0',
+            'b/level_1/model/qf_1/fc1/bias:0',
+            'b/level_1/model/qf_1/fc1/kernel:0',
+            'b/level_1/model/qf_1/qf_output/bias:0',
+            'b/level_1/model/qf_1/qf_output/kernel:0',
+        ]
+
+        target_var_list = [
+            'a/level_0/target/pi/fc0/bias:0',
+            'a/level_0/target/pi/fc0/kernel:0',
+            'a/level_0/target/pi/fc1/bias:0',
+            'a/level_0/target/pi/fc1/kernel:0',
+            'a/level_0/target/pi/output/bias:0',
+            'a/level_0/target/pi/output/kernel:0',
+            'a/level_0/target/qf_0/fc0/bias:0',
+            'a/level_0/target/qf_0/fc0/kernel:0',
+            'a/level_0/target/qf_0/fc1/bias:0',
+            'a/level_0/target/qf_0/fc1/kernel:0',
+            'a/level_0/target/qf_0/qf_output/bias:0',
+            'a/level_0/target/qf_0/qf_output/kernel:0',
+            'a/level_0/target/qf_1/fc0/bias:0',
+            'a/level_0/target/qf_1/fc0/kernel:0',
+            'a/level_0/target/qf_1/fc1/bias:0',
+            'a/level_0/target/qf_1/fc1/kernel:0',
+            'a/level_0/target/qf_1/qf_output/bias:0',
+            'a/level_0/target/qf_1/qf_output/kernel:0',
+
+            'a/level_1/target/pi/fc0/bias:0',
+            'a/level_1/target/pi/fc0/kernel:0',
+            'a/level_1/target/pi/fc1/bias:0',
+            'a/level_1/target/pi/fc1/kernel:0',
+            'a/level_1/target/pi/output/bias:0',
+            'a/level_1/target/pi/output/kernel:0',
+            'a/level_1/target/qf_0/fc0/bias:0',
+            'a/level_1/target/qf_0/fc0/kernel:0',
+            'a/level_1/target/qf_0/fc1/bias:0',
+            'a/level_1/target/qf_0/fc1/kernel:0',
+            'a/level_1/target/qf_0/qf_output/bias:0',
+            'a/level_1/target/qf_0/qf_output/kernel:0',
+            'a/level_1/target/qf_1/fc0/bias:0',
+            'a/level_1/target/qf_1/fc0/kernel:0',
+            'a/level_1/target/qf_1/fc1/bias:0',
+            'a/level_1/target/qf_1/fc1/kernel:0',
+            'a/level_1/target/qf_1/qf_output/bias:0',
+            'a/level_1/target/qf_1/qf_output/kernel:0',
+
+            'b/level_0/target/pi/fc0/bias:0',
+            'b/level_0/target/pi/fc0/kernel:0',
+            'b/level_0/target/pi/fc1/bias:0',
+            'b/level_0/target/pi/fc1/kernel:0',
+            'b/level_0/target/pi/output/bias:0',
+            'b/level_0/target/pi/output/kernel:0',
+            'b/level_0/target/qf_0/fc0/bias:0',
+            'b/level_0/target/qf_0/fc0/kernel:0',
+            'b/level_0/target/qf_0/fc1/bias:0',
+            'b/level_0/target/qf_0/fc1/kernel:0',
+            'b/level_0/target/qf_0/qf_output/bias:0',
+            'b/level_0/target/qf_0/qf_output/kernel:0',
+            'b/level_0/target/qf_1/fc0/bias:0',
+            'b/level_0/target/qf_1/fc0/kernel:0',
+            'b/level_0/target/qf_1/fc1/bias:0',
+            'b/level_0/target/qf_1/fc1/kernel:0',
+            'b/level_0/target/qf_1/qf_output/bias:0',
+            'b/level_0/target/qf_1/qf_output/kernel:0',
+
+            'b/level_1/target/pi/fc0/bias:0',
+            'b/level_1/target/pi/fc0/kernel:0',
+            'b/level_1/target/pi/fc1/bias:0',
+            'b/level_1/target/pi/fc1/kernel:0',
+            'b/level_1/target/pi/output/bias:0',
+            'b/level_1/target/pi/output/kernel:0',
+            'b/level_1/target/qf_0/fc0/bias:0',
+            'b/level_1/target/qf_0/fc0/kernel:0',
+            'b/level_1/target/qf_0/fc1/bias:0',
+            'b/level_1/target/qf_0/fc1/kernel:0',
+            'b/level_1/target/qf_0/qf_output/bias:0',
+            'b/level_1/target/qf_0/qf_output/kernel:0',
+            'b/level_1/target/qf_1/fc0/bias:0',
+            'b/level_1/target/qf_1/fc0/kernel:0',
+            'b/level_1/target/qf_1/fc1/bias:0',
+            'b/level_1/target/qf_1/fc1/kernel:0',
+            'b/level_1/target/qf_1/qf_output/bias:0',
+            'b/level_1/target/qf_1/qf_output/kernel:0',
+        ]
+
+        for model, target in zip(model_var_list, target_var_list):
+            with tf.compat.v1.variable_scope(
+                    tf.compat.v1.get_variable_scope(), reuse=True):
+                model_val = policy.sess.run(model)
+                target_val = policy.sess.run(target)
+            np.testing.assert_almost_equal(model_val, target_val)
+
+    def test_initialize_2(self):
+        policy_params = self.policy_params_shared.copy()
+        policy_params["maddpg"] = False
+        policy = TD3MultiGoalConditionedPolicy(**policy_params)
+
+        # Initialize the variables of the policy.
+        policy.sess.run(tf.compat.v1.global_variables_initializer())
+
+        # Run the initialize method.
+        policy.initialize()
+
+        model_var_list = [
+            'level_0/model/pi/fc0/bias:0',
+            'level_0/model/pi/fc0/kernel:0',
+            'level_0/model/pi/fc1/bias:0',
+            'level_0/model/pi/fc1/kernel:0',
+            'level_0/model/pi/output/bias:0',
+            'level_0/model/pi/output/kernel:0',
+            'level_0/model/qf_0/fc0/bias:0',
+            'level_0/model/qf_0/fc0/kernel:0',
+            'level_0/model/qf_0/fc1/bias:0',
+            'level_0/model/qf_0/fc1/kernel:0',
+            'level_0/model/qf_0/qf_output/bias:0',
+            'level_0/model/qf_0/qf_output/kernel:0',
+            'level_0/model/qf_1/fc0/bias:0',
+            'level_0/model/qf_1/fc0/kernel:0',
+            'level_0/model/qf_1/fc1/bias:0',
+            'level_0/model/qf_1/fc1/kernel:0',
+            'level_0/model/qf_1/qf_output/bias:0',
+            'level_0/model/qf_1/qf_output/kernel:0',
+
+            'level_1/model/pi/fc0/bias:0',
+            'level_1/model/pi/fc0/kernel:0',
+            'level_1/model/pi/fc1/bias:0',
+            'level_1/model/pi/fc1/kernel:0',
+            'level_1/model/pi/output/bias:0',
+            'level_1/model/pi/output/kernel:0',
+            'level_1/model/qf_0/fc0/bias:0',
+            'level_1/model/qf_0/fc0/kernel:0',
+            'level_1/model/qf_0/fc1/bias:0',
+            'level_1/model/qf_0/fc1/kernel:0',
+            'level_1/model/qf_0/qf_output/bias:0',
+            'level_1/model/qf_0/qf_output/kernel:0',
+            'level_1/model/qf_1/fc0/bias:0',
+            'level_1/model/qf_1/fc0/kernel:0',
+            'level_1/model/qf_1/fc1/bias:0',
+            'level_1/model/qf_1/fc1/kernel:0',
+            'level_1/model/qf_1/qf_output/bias:0',
+            'level_1/model/qf_1/qf_output/kernel:0',
+        ]
+
+        target_var_list = [
+            'level_0/target/pi/fc0/bias:0',
+            'level_0/target/pi/fc0/kernel:0',
+            'level_0/target/pi/fc1/bias:0',
+            'level_0/target/pi/fc1/kernel:0',
+            'level_0/target/pi/output/bias:0',
+            'level_0/target/pi/output/kernel:0',
+            'level_0/target/qf_0/fc0/bias:0',
+            'level_0/target/qf_0/fc0/kernel:0',
+            'level_0/target/qf_0/fc1/bias:0',
+            'level_0/target/qf_0/fc1/kernel:0',
+            'level_0/target/qf_0/qf_output/bias:0',
+            'level_0/target/qf_0/qf_output/kernel:0',
+            'level_0/target/qf_1/fc0/bias:0',
+            'level_0/target/qf_1/fc0/kernel:0',
+            'level_0/target/qf_1/fc1/bias:0',
+            'level_0/target/qf_1/fc1/kernel:0',
+            'level_0/target/qf_1/qf_output/bias:0',
+            'level_0/target/qf_1/qf_output/kernel:0',
+
+            'level_1/target/pi/fc0/bias:0',
+            'level_1/target/pi/fc0/kernel:0',
+            'level_1/target/pi/fc1/bias:0',
+            'level_1/target/pi/fc1/kernel:0',
+            'level_1/target/pi/output/bias:0',
+            'level_1/target/pi/output/kernel:0',
+            'level_1/target/qf_0/fc0/bias:0',
+            'level_1/target/qf_0/fc0/kernel:0',
+            'level_1/target/qf_0/fc1/bias:0',
+            'level_1/target/qf_0/fc1/kernel:0',
+            'level_1/target/qf_0/qf_output/bias:0',
+            'level_1/target/qf_0/qf_output/kernel:0',
+            'level_1/target/qf_1/fc0/bias:0',
+            'level_1/target/qf_1/fc0/kernel:0',
+            'level_1/target/qf_1/fc1/bias:0',
+            'level_1/target/qf_1/fc1/kernel:0',
+            'level_1/target/qf_1/qf_output/bias:0',
+            'level_1/target/qf_1/qf_output/kernel:0',
+        ]
+
+        for model, target in zip(model_var_list, target_var_list):
+            with tf.compat.v1.variable_scope(
+                    tf.compat.v1.get_variable_scope(), reuse=True):
+                model_val = policy.sess.run(model)
+                target_val = policy.sess.run(target)
+            np.testing.assert_almost_equal(model_val, target_val)
+
+
+class TestSACMultiGoalConditionedPolicy(unittest.TestCase):
+    """Test MultiFeedForwardPolicy in hbaselines/multiagent/h_sac.py."""
+
+    def setUp(self):
+        self.sess = tf.compat.v1.Session()
+
+        # Shared policy parameters
+        self.policy_params_shared = {
+            'sess': self.sess,
+            'ac_space': Box(low=-1, high=1, shape=(1,)),
+            'co_space': Box(low=-2, high=2, shape=(2,)),
+            'ob_space': Box(low=-3, high=3, shape=(3,)),
+            'all_ob_space': Box(low=-3, high=3, shape=(10,)),
+            'layers': [256, 256],
+            'verbose': 0,
+        }
+        self.policy_params_shared.update(SAC_PARAMS.copy())
+        self.policy_params_shared.update(GOAL_CONDITIONED_PARAMS.copy())
+        self.policy_params_shared.update(MULTIAGENT_PARAMS.copy())
+        self.policy_params_shared['shared'] = True
+
+        # Independent policy parameters
+        self.policy_params_independent = {
+            'sess': self.sess,
+            'ac_space': {
+                'a': Box(low=-1, high=1, shape=(1,)),
+                'b': Box(low=-2, high=2, shape=(2,)),
+            },
+            'co_space': {
+                'a': Box(low=-3, high=3, shape=(3,)),
+                'b': Box(low=-4, high=4, shape=(4,)),
+            },
+            'ob_space': {
+                'a': Box(low=-5, high=5, shape=(5,)),
+                'b': Box(low=-6, high=6, shape=(6,)),
+            },
+            'all_ob_space': Box(low=-6, high=6, shape=(18,)),
+            'layers': [256, 256],
+            'verbose': 0,
+        }
+        self.policy_params_independent.update(SAC_PARAMS.copy())
+        self.policy_params_independent.update(GOAL_CONDITIONED_PARAMS.copy())
+        self.policy_params_independent.update(MULTIAGENT_PARAMS.copy())
+        self.policy_params_independent['shared'] = False
+
+    def tearDown(self):
+        self.sess.close()
+        del self.policy_params_shared
+        del self.policy_params_independent
+
+        # Clear the graph.
+        tf.compat.v1.reset_default_graph()
+
+    def test_init_1(self):
+        """Check the functionality of the __init__() method.
+
+        This method is tested for the following features:
+
+        1. The proper structure graph was generated.
+        2. All input placeholders are correct.
+
+        This is done for the following cases:
+
+        1. maddpg = False, shared = False
+        2. maddpg = False, shared = True
+        """
+        policy_params = self.policy_params_independent.copy()
+        policy_params["maddpg"] = False
+        policy = SACMultiGoalConditionedPolicy(**policy_params)
+
+        self.assertListEqual(
+            sorted([var.name for var in get_trainable_vars()]),
+            ['a/level_0/model/log_alpha:0',
+             'a/level_0/model/pi/fc0/bias:0',
+             'a/level_0/model/pi/fc0/kernel:0',
+             'a/level_0/model/pi/fc1/bias:0',
+             'a/level_0/model/pi/fc1/kernel:0',
+             'a/level_0/model/pi/log_std/bias:0',
+             'a/level_0/model/pi/log_std/kernel:0',
+             'a/level_0/model/pi/mean/bias:0',
+             'a/level_0/model/pi/mean/kernel:0',
+             'a/level_0/model/value_fns/qf1/fc0/bias:0',
+             'a/level_0/model/value_fns/qf1/fc0/kernel:0',
+             'a/level_0/model/value_fns/qf1/fc1/bias:0',
+             'a/level_0/model/value_fns/qf1/fc1/kernel:0',
+             'a/level_0/model/value_fns/qf1/qf_output/bias:0',
+             'a/level_0/model/value_fns/qf1/qf_output/kernel:0',
+             'a/level_0/model/value_fns/qf2/fc0/bias:0',
+             'a/level_0/model/value_fns/qf2/fc0/kernel:0',
+             'a/level_0/model/value_fns/qf2/fc1/bias:0',
+             'a/level_0/model/value_fns/qf2/fc1/kernel:0',
+             'a/level_0/model/value_fns/qf2/qf_output/bias:0',
+             'a/level_0/model/value_fns/qf2/qf_output/kernel:0',
+             'a/level_0/model/value_fns/vf/fc0/bias:0',
+             'a/level_0/model/value_fns/vf/fc0/kernel:0',
+             'a/level_0/model/value_fns/vf/fc1/bias:0',
+             'a/level_0/model/value_fns/vf/fc1/kernel:0',
+             'a/level_0/model/value_fns/vf/vf_output/bias:0',
+             'a/level_0/model/value_fns/vf/vf_output/kernel:0',
+             'a/level_0/target/value_fns/vf/fc0/bias:0',
+             'a/level_0/target/value_fns/vf/fc0/kernel:0',
+             'a/level_0/target/value_fns/vf/fc1/bias:0',
+             'a/level_0/target/value_fns/vf/fc1/kernel:0',
+             'a/level_0/target/value_fns/vf/vf_output/bias:0',
+             'a/level_0/target/value_fns/vf/vf_output/kernel:0',
+             'a/level_1/model/log_alpha:0',
+             'a/level_1/model/pi/fc0/bias:0',
+             'a/level_1/model/pi/fc0/kernel:0',
+             'a/level_1/model/pi/fc1/bias:0',
+             'a/level_1/model/pi/fc1/kernel:0',
+             'a/level_1/model/pi/log_std/bias:0',
+             'a/level_1/model/pi/log_std/kernel:0',
+             'a/level_1/model/pi/mean/bias:0',
+             'a/level_1/model/pi/mean/kernel:0',
+             'a/level_1/model/value_fns/qf1/fc0/bias:0',
+             'a/level_1/model/value_fns/qf1/fc0/kernel:0',
+             'a/level_1/model/value_fns/qf1/fc1/bias:0',
+             'a/level_1/model/value_fns/qf1/fc1/kernel:0',
+             'a/level_1/model/value_fns/qf1/qf_output/bias:0',
+             'a/level_1/model/value_fns/qf1/qf_output/kernel:0',
+             'a/level_1/model/value_fns/qf2/fc0/bias:0',
+             'a/level_1/model/value_fns/qf2/fc0/kernel:0',
+             'a/level_1/model/value_fns/qf2/fc1/bias:0',
+             'a/level_1/model/value_fns/qf2/fc1/kernel:0',
+             'a/level_1/model/value_fns/qf2/qf_output/bias:0',
+             'a/level_1/model/value_fns/qf2/qf_output/kernel:0',
+             'a/level_1/model/value_fns/vf/fc0/bias:0',
+             'a/level_1/model/value_fns/vf/fc0/kernel:0',
+             'a/level_1/model/value_fns/vf/fc1/bias:0',
+             'a/level_1/model/value_fns/vf/fc1/kernel:0',
+             'a/level_1/model/value_fns/vf/vf_output/bias:0',
+             'a/level_1/model/value_fns/vf/vf_output/kernel:0',
+             'a/level_1/target/value_fns/vf/fc0/bias:0',
+             'a/level_1/target/value_fns/vf/fc0/kernel:0',
+             'a/level_1/target/value_fns/vf/fc1/bias:0',
+             'a/level_1/target/value_fns/vf/fc1/kernel:0',
+             'a/level_1/target/value_fns/vf/vf_output/bias:0',
+             'a/level_1/target/value_fns/vf/vf_output/kernel:0',
+             'b/level_0/model/log_alpha:0',
+             'b/level_0/model/pi/fc0/bias:0',
+             'b/level_0/model/pi/fc0/kernel:0',
+             'b/level_0/model/pi/fc1/bias:0',
+             'b/level_0/model/pi/fc1/kernel:0',
+             'b/level_0/model/pi/log_std/bias:0',
+             'b/level_0/model/pi/log_std/kernel:0',
+             'b/level_0/model/pi/mean/bias:0',
+             'b/level_0/model/pi/mean/kernel:0',
+             'b/level_0/model/value_fns/qf1/fc0/bias:0',
+             'b/level_0/model/value_fns/qf1/fc0/kernel:0',
+             'b/level_0/model/value_fns/qf1/fc1/bias:0',
+             'b/level_0/model/value_fns/qf1/fc1/kernel:0',
+             'b/level_0/model/value_fns/qf1/qf_output/bias:0',
+             'b/level_0/model/value_fns/qf1/qf_output/kernel:0',
+             'b/level_0/model/value_fns/qf2/fc0/bias:0',
+             'b/level_0/model/value_fns/qf2/fc0/kernel:0',
+             'b/level_0/model/value_fns/qf2/fc1/bias:0',
+             'b/level_0/model/value_fns/qf2/fc1/kernel:0',
+             'b/level_0/model/value_fns/qf2/qf_output/bias:0',
+             'b/level_0/model/value_fns/qf2/qf_output/kernel:0',
+             'b/level_0/model/value_fns/vf/fc0/bias:0',
+             'b/level_0/model/value_fns/vf/fc0/kernel:0',
+             'b/level_0/model/value_fns/vf/fc1/bias:0',
+             'b/level_0/model/value_fns/vf/fc1/kernel:0',
+             'b/level_0/model/value_fns/vf/vf_output/bias:0',
+             'b/level_0/model/value_fns/vf/vf_output/kernel:0',
+             'b/level_0/target/value_fns/vf/fc0/bias:0',
+             'b/level_0/target/value_fns/vf/fc0/kernel:0',
+             'b/level_0/target/value_fns/vf/fc1/bias:0',
+             'b/level_0/target/value_fns/vf/fc1/kernel:0',
+             'b/level_0/target/value_fns/vf/vf_output/bias:0',
+             'b/level_0/target/value_fns/vf/vf_output/kernel:0',
+             'b/level_1/model/log_alpha:0',
+             'b/level_1/model/pi/fc0/bias:0',
+             'b/level_1/model/pi/fc0/kernel:0',
+             'b/level_1/model/pi/fc1/bias:0',
+             'b/level_1/model/pi/fc1/kernel:0',
+             'b/level_1/model/pi/log_std/bias:0',
+             'b/level_1/model/pi/log_std/kernel:0',
+             'b/level_1/model/pi/mean/bias:0',
+             'b/level_1/model/pi/mean/kernel:0',
+             'b/level_1/model/value_fns/qf1/fc0/bias:0',
+             'b/level_1/model/value_fns/qf1/fc0/kernel:0',
+             'b/level_1/model/value_fns/qf1/fc1/bias:0',
+             'b/level_1/model/value_fns/qf1/fc1/kernel:0',
+             'b/level_1/model/value_fns/qf1/qf_output/bias:0',
+             'b/level_1/model/value_fns/qf1/qf_output/kernel:0',
+             'b/level_1/model/value_fns/qf2/fc0/bias:0',
+             'b/level_1/model/value_fns/qf2/fc0/kernel:0',
+             'b/level_1/model/value_fns/qf2/fc1/bias:0',
+             'b/level_1/model/value_fns/qf2/fc1/kernel:0',
+             'b/level_1/model/value_fns/qf2/qf_output/bias:0',
+             'b/level_1/model/value_fns/qf2/qf_output/kernel:0',
+             'b/level_1/model/value_fns/vf/fc0/bias:0',
+             'b/level_1/model/value_fns/vf/fc0/kernel:0',
+             'b/level_1/model/value_fns/vf/fc1/bias:0',
+             'b/level_1/model/value_fns/vf/fc1/kernel:0',
+             'b/level_1/model/value_fns/vf/vf_output/bias:0',
+             'b/level_1/model/value_fns/vf/vf_output/kernel:0',
+             'b/level_1/target/value_fns/vf/fc0/bias:0',
+             'b/level_1/target/value_fns/vf/fc0/kernel:0',
+             'b/level_1/target/value_fns/vf/fc1/bias:0',
+             'b/level_1/target/value_fns/vf/fc1/kernel:0',
+             'b/level_1/target/value_fns/vf/vf_output/bias:0',
+             'b/level_1/target/value_fns/vf/vf_output/kernel:0']
+        )
+
+        # Check observation/action/context spaces of the agents
+        self.assertEqual(
+            policy.agents['a'].policy[0].ac_space.shape[0],
+            self.policy_params_independent['ob_space']['a'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['a'].policy[0].ob_space.shape[0],
+            self.policy_params_independent['ob_space']['a'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['a'].policy[0].co_space.shape[0],
+            self.policy_params_independent['co_space']['a'].shape[0]
+        )
+
+        self.assertEqual(
+            policy.agents['a'].policy[1].ac_space.shape[0],
+            self.policy_params_independent['ac_space']['a'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['a'].policy[1].ob_space.shape[0],
+            self.policy_params_independent['ob_space']['a'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['a'].policy[1].co_space.shape[0],
+            self.policy_params_independent['ob_space']['a'].shape[0]
+        )
+
+        self.assertEqual(
+            policy.agents['b'].policy[0].ac_space.shape[0],
+            self.policy_params_independent['ob_space']['b'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['b'].policy[0].ob_space.shape[0],
+            self.policy_params_independent['ob_space']['b'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['b'].policy[0].co_space.shape[0],
+            self.policy_params_independent['co_space']['b'].shape[0]
+        )
+
+        self.assertEqual(
+            policy.agents['b'].policy[1].ac_space.shape[0],
+            self.policy_params_independent['ac_space']['b'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['b'].policy[1].ob_space.shape[0],
+            self.policy_params_independent['ob_space']['b'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['b'].policy[1].co_space.shape[0],
+            self.policy_params_independent['ob_space']['b'].shape[0]
+        )
+
+        # Check the instantiation of the class attributes.
+        self.assertTrue(not policy.shared)
+        self.assertTrue(not policy.maddpg)
+
+    def test_init_2(self):
+        policy_params = self.policy_params_shared.copy()
+        policy_params["maddpg"] = False
+        policy = SACMultiGoalConditionedPolicy(**policy_params)
+
+        self.assertListEqual(
+            sorted([var.name for var in get_trainable_vars()]),
+            ['level_0/model/log_alpha:0',
+             'level_0/model/pi/fc0/bias:0',
+             'level_0/model/pi/fc0/kernel:0',
+             'level_0/model/pi/fc1/bias:0',
+             'level_0/model/pi/fc1/kernel:0',
+             'level_0/model/pi/log_std/bias:0',
+             'level_0/model/pi/log_std/kernel:0',
+             'level_0/model/pi/mean/bias:0',
+             'level_0/model/pi/mean/kernel:0',
+             'level_0/model/value_fns/qf1/fc0/bias:0',
+             'level_0/model/value_fns/qf1/fc0/kernel:0',
+             'level_0/model/value_fns/qf1/fc1/bias:0',
+             'level_0/model/value_fns/qf1/fc1/kernel:0',
+             'level_0/model/value_fns/qf1/qf_output/bias:0',
+             'level_0/model/value_fns/qf1/qf_output/kernel:0',
+             'level_0/model/value_fns/qf2/fc0/bias:0',
+             'level_0/model/value_fns/qf2/fc0/kernel:0',
+             'level_0/model/value_fns/qf2/fc1/bias:0',
+             'level_0/model/value_fns/qf2/fc1/kernel:0',
+             'level_0/model/value_fns/qf2/qf_output/bias:0',
+             'level_0/model/value_fns/qf2/qf_output/kernel:0',
+             'level_0/model/value_fns/vf/fc0/bias:0',
+             'level_0/model/value_fns/vf/fc0/kernel:0',
+             'level_0/model/value_fns/vf/fc1/bias:0',
+             'level_0/model/value_fns/vf/fc1/kernel:0',
+             'level_0/model/value_fns/vf/vf_output/bias:0',
+             'level_0/model/value_fns/vf/vf_output/kernel:0',
+             'level_0/target/value_fns/vf/fc0/bias:0',
+             'level_0/target/value_fns/vf/fc0/kernel:0',
+             'level_0/target/value_fns/vf/fc1/bias:0',
+             'level_0/target/value_fns/vf/fc1/kernel:0',
+             'level_0/target/value_fns/vf/vf_output/bias:0',
+             'level_0/target/value_fns/vf/vf_output/kernel:0',
+             'level_1/model/log_alpha:0',
+             'level_1/model/pi/fc0/bias:0',
+             'level_1/model/pi/fc0/kernel:0',
+             'level_1/model/pi/fc1/bias:0',
+             'level_1/model/pi/fc1/kernel:0',
+             'level_1/model/pi/log_std/bias:0',
+             'level_1/model/pi/log_std/kernel:0',
+             'level_1/model/pi/mean/bias:0',
+             'level_1/model/pi/mean/kernel:0',
+             'level_1/model/value_fns/qf1/fc0/bias:0',
+             'level_1/model/value_fns/qf1/fc0/kernel:0',
+             'level_1/model/value_fns/qf1/fc1/bias:0',
+             'level_1/model/value_fns/qf1/fc1/kernel:0',
+             'level_1/model/value_fns/qf1/qf_output/bias:0',
+             'level_1/model/value_fns/qf1/qf_output/kernel:0',
+             'level_1/model/value_fns/qf2/fc0/bias:0',
+             'level_1/model/value_fns/qf2/fc0/kernel:0',
+             'level_1/model/value_fns/qf2/fc1/bias:0',
+             'level_1/model/value_fns/qf2/fc1/kernel:0',
+             'level_1/model/value_fns/qf2/qf_output/bias:0',
+             'level_1/model/value_fns/qf2/qf_output/kernel:0',
+             'level_1/model/value_fns/vf/fc0/bias:0',
+             'level_1/model/value_fns/vf/fc0/kernel:0',
+             'level_1/model/value_fns/vf/fc1/bias:0',
+             'level_1/model/value_fns/vf/fc1/kernel:0',
+             'level_1/model/value_fns/vf/vf_output/bias:0',
+             'level_1/model/value_fns/vf/vf_output/kernel:0',
+             'level_1/target/value_fns/vf/fc0/bias:0',
+             'level_1/target/value_fns/vf/fc0/kernel:0',
+             'level_1/target/value_fns/vf/fc1/bias:0',
+             'level_1/target/value_fns/vf/fc1/kernel:0',
+             'level_1/target/value_fns/vf/vf_output/bias:0',
+             'level_1/target/value_fns/vf/vf_output/kernel:0']
+        )
+
+        # Check observation/action/context spaces of the agents
+        self.assertEqual(
+            policy.agents['policy'].policy[0].ac_space.shape[0],
+            self.policy_params_shared['ob_space'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['policy'].policy[0].ob_space.shape[0],
+            self.policy_params_shared['ob_space'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['policy'].policy[0].co_space.shape[0],
+            self.policy_params_shared['co_space'].shape[0]
+        )
+
+        self.assertEqual(
+            policy.agents['policy'].policy[1].ac_space.shape[0],
+            self.policy_params_shared['ac_space'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['policy'].policy[1].ob_space.shape[0],
+            self.policy_params_shared['ob_space'].shape[0]
+        )
+        self.assertEqual(
+            policy.agents['policy'].policy[1].co_space.shape[0],
+            self.policy_params_shared['ob_space'].shape[0]
+        )
+
+        # Check the instantiation of the class attributes.
+        self.assertTrue(policy.shared)
+        self.assertTrue(not policy.maddpg)
+
+    def test_initialize_1(self):
+        """Check the functionality of the initialize() method.
+
+        This test validates that the target variables are properly initialized
+        when initialize is called.
+
+        This is done for the following cases:
+
+        1. maddpg = False, shared = False
+        2. maddpg = False, shared = True
+        """
+        policy_params = self.policy_params_independent.copy()
+        policy_params["maddpg"] = False
+        policy = SACMultiGoalConditionedPolicy(**policy_params)
+
+        # Initialize the variables of the policy.
+        policy.sess.run(tf.compat.v1.global_variables_initializer())
+
+        # Run the initialize method.
+        policy.initialize()
+
+        model_var_list = [
+            'a/level_0/model/value_fns/vf/fc0/kernel:0',
+            'a/level_0/model/value_fns/vf/fc0/bias:0',
+            'a/level_0/model/value_fns/vf/fc1/kernel:0',
+            'a/level_0/model/value_fns/vf/fc1/bias:0',
+            'a/level_0/model/value_fns/vf/vf_output/kernel:0',
+            'a/level_0/model/value_fns/vf/vf_output/bias:0',
+
+            'a/level_1/model/value_fns/vf/fc0/kernel:0',
+            'a/level_1/model/value_fns/vf/fc0/bias:0',
+            'a/level_1/model/value_fns/vf/fc1/kernel:0',
+            'a/level_1/model/value_fns/vf/fc1/bias:0',
+            'a/level_1/model/value_fns/vf/vf_output/kernel:0',
+            'a/level_1/model/value_fns/vf/vf_output/bias:0',
+
+            'b/level_0/model/value_fns/vf/fc0/kernel:0',
+            'b/level_0/model/value_fns/vf/fc0/bias:0',
+            'b/level_0/model/value_fns/vf/fc1/kernel:0',
+            'b/level_0/model/value_fns/vf/fc1/bias:0',
+            'b/level_0/model/value_fns/vf/vf_output/kernel:0',
+            'b/level_0/model/value_fns/vf/vf_output/bias:0',
+
+            'b/level_1/model/value_fns/vf/fc0/kernel:0',
+            'b/level_1/model/value_fns/vf/fc0/bias:0',
+            'b/level_1/model/value_fns/vf/fc1/kernel:0',
+            'b/level_1/model/value_fns/vf/fc1/bias:0',
+            'b/level_1/model/value_fns/vf/vf_output/kernel:0',
+            'b/level_1/model/value_fns/vf/vf_output/bias:0',
+        ]
+
+        target_var_list = [
+            'a/level_0/target/value_fns/vf/fc0/kernel:0',
+            'a/level_0/target/value_fns/vf/fc0/bias:0',
+            'a/level_0/target/value_fns/vf/fc1/kernel:0',
+            'a/level_0/target/value_fns/vf/fc1/bias:0',
+            'a/level_0/target/value_fns/vf/vf_output/kernel:0',
+            'a/level_0/target/value_fns/vf/vf_output/bias:0',
+
+            'a/level_1/target/value_fns/vf/fc0/kernel:0',
+            'a/level_1/target/value_fns/vf/fc0/bias:0',
+            'a/level_1/target/value_fns/vf/fc1/kernel:0',
+            'a/level_1/target/value_fns/vf/fc1/bias:0',
+            'a/level_1/target/value_fns/vf/vf_output/kernel:0',
+            'a/level_1/target/value_fns/vf/vf_output/bias:0',
+
+            'b/level_0/target/value_fns/vf/fc0/kernel:0',
+            'b/level_0/target/value_fns/vf/fc0/bias:0',
+            'b/level_0/target/value_fns/vf/fc1/kernel:0',
+            'b/level_0/target/value_fns/vf/fc1/bias:0',
+            'b/level_0/target/value_fns/vf/vf_output/kernel:0',
+            'b/level_0/target/value_fns/vf/vf_output/bias:0',
+
+            'b/level_1/target/value_fns/vf/fc0/kernel:0',
+            'b/level_1/target/value_fns/vf/fc0/bias:0',
+            'b/level_1/target/value_fns/vf/fc1/kernel:0',
+            'b/level_1/target/value_fns/vf/fc1/bias:0',
+            'b/level_1/target/value_fns/vf/vf_output/kernel:0',
+            'b/level_1/target/value_fns/vf/vf_output/bias:0',
+        ]
+
+        for model, target in zip(model_var_list, target_var_list):
+            with tf.compat.v1.variable_scope(
+                    tf.compat.v1.get_variable_scope(), reuse=True):
+                model_val = policy.sess.run(model)
+                target_val = policy.sess.run(target)
+            np.testing.assert_almost_equal(model_val, target_val)
+
+    def test_initialize_2(self):
+        policy_params = self.policy_params_shared.copy()
+        policy_params["maddpg"] = False
+        policy = SACMultiGoalConditionedPolicy(**policy_params)
+
+        # Initialize the variables of the policy.
+        policy.sess.run(tf.compat.v1.global_variables_initializer())
+
+        # Run the initialize method.
+        policy.initialize()
+
+        model_var_list = [
+            'level_0/model/value_fns/vf/fc0/kernel:0',
+            'level_0/model/value_fns/vf/fc0/bias:0',
+            'level_0/model/value_fns/vf/fc1/kernel:0',
+            'level_0/model/value_fns/vf/fc1/bias:0',
+            'level_0/model/value_fns/vf/vf_output/kernel:0',
+            'level_0/model/value_fns/vf/vf_output/bias:0',
+
+            'level_1/model/value_fns/vf/fc0/kernel:0',
+            'level_1/model/value_fns/vf/fc0/bias:0',
+            'level_1/model/value_fns/vf/fc1/kernel:0',
+            'level_1/model/value_fns/vf/fc1/bias:0',
+            'level_1/model/value_fns/vf/vf_output/kernel:0',
+            'level_1/model/value_fns/vf/vf_output/bias:0',
+        ]
+
+        target_var_list = [
+            'level_0/target/value_fns/vf/fc0/kernel:0',
+            'level_0/target/value_fns/vf/fc0/bias:0',
+            'level_0/target/value_fns/vf/fc1/kernel:0',
+            'level_0/target/value_fns/vf/fc1/bias:0',
+            'level_0/target/value_fns/vf/vf_output/kernel:0',
+            'level_0/target/value_fns/vf/vf_output/bias:0',
+
+            'level_1/target/value_fns/vf/fc0/kernel:0',
+            'level_1/target/value_fns/vf/fc0/bias:0',
+            'level_1/target/value_fns/vf/fc1/kernel:0',
+            'level_1/target/value_fns/vf/fc1/bias:0',
+            'level_1/target/value_fns/vf/vf_output/kernel:0',
+            'level_1/target/value_fns/vf/vf_output/bias:0',
         ]
 
         for model, target in zip(model_var_list, target_var_list):
