+++ tests/slow_tests/test_experiments.py
@@ -12,6 +12,7 @@ from experiments.plot import plot_fig
 from experiments.run_fcnet import main as run_fcnet
 from experiments.run_hrl import main as run_hrl
 from experiments.run_multi_fcnet import main as run_multi_fcnet
+from experiments.run_multi_hrl import main as run_multi_hrl
 
 os.environ["TEST_FLAG"] = "True"
 
@@ -465,6 +466,119 @@ class TestExperimentRunnerScripts(unittest.TestCase):
 
         self.assertRaises(ValueError, run_multi_fcnet,
                           args=args, base_dir='data/multi-fcnet')
+
+        # Clear anything that was generated.
+        shutil.rmtree(os.path.join(os.getcwd(), "data"))
+
+    def test_run_multi_hrl_td3_independent(self):
+        # Run the script; verify it executes without failure.
+        args = parse_train_options('', '', args=[
+            "multiagent-ring_small",
+            "--initial_exploration_steps", "1",
+            "--total_steps", "500",
+            "--log_interval", "500",
+        ])
+        run_multi_hrl(args, 'data/multi-goal-conditioned')
+
+        # Check that the folders were generated.
+        self.assertTrue(os.path.isdir(
+            os.path.join(
+                os.getcwd(),
+                "data/multi-goal-conditioned/multiagent-ring_small")))
+
+        # Clear anything that was generated.
+        shutil.rmtree(os.path.join(os.getcwd(), "data"))
+
+    def test_run_multi_hrl_sac_independent(self):
+        # Run the script; verify it executes without failure.
+        args = parse_train_options('', '', args=[
+            "multiagent-ring_small",
+            "--initial_exploration_steps", "1",
+            "--total_steps", "500",
+            "--log_interval", "500",
+            "--alg", "SAC"
+        ])
+        run_multi_hrl(args, 'data/multi-goal-conditioned')
+
+        # Check that the folders were generated.
+        self.assertTrue(os.path.isdir(
+            os.path.join(
+                os.getcwd(),
+                "data/multi-goal-conditioned/multiagent-ring_small")))
+
+        # Clear anything that was generated.
+        shutil.rmtree(os.path.join(os.getcwd(), "data"))
+
+    def test_run_multi_hrl_failure_independent(self):
+        # Run the script; verify it executes without failure.
+        args = parse_train_options('', '', args=[
+            "multiagent-ring_small",
+            "--initial_exploration_steps", "1",
+            "--total_steps", "500",
+            "--log_interval", "500",
+            "--alg", "woops"
+        ])
+
+        self.assertRaises(ValueError, run_multi_hrl,
+                          args=args, base_dir='data/multi-goal-conditioned')
+
+        # Clear anything that was generated.
+        shutil.rmtree(os.path.join(os.getcwd(), "data"))
+
+    def test_run_multi_hrl_td3_shared(self):
+        # Run the script; verify it executes without failure.
+        args = parse_train_options('', '', args=[
+            "multiagent-ring_small",
+            "--shared",
+            "--initial_exploration_steps", "1",
+            "--total_steps", "500",
+            "--log_interval", "500",
+        ])
+        run_multi_hrl(args, 'data/multi-goal-conditioned')
+
+        # Check that the folders were generated.
+        self.assertTrue(os.path.isdir(
+            os.path.join(
+                os.getcwd(),
+                "data/multi-goal-conditioned/multiagent-ring_small")))
+
+        # Clear anything that was generated.
+        shutil.rmtree(os.path.join(os.getcwd(), "data"))
+
+    def test_run_multi_hrl_sac_shared(self):
+        # Run the script; verify it executes without failure.
+        args = parse_train_options('', '', args=[
+            "multiagent-ring_small",
+            "--shared",
+            "--initial_exploration_steps", "1",
+            "--total_steps", "500",
+            "--log_interval", "500",
+            "--alg", "SAC"
+        ])
+        run_multi_hrl(args, 'data/multi-goal-conditioned')
+
+        # Check that the folders were generated.
+        self.assertTrue(os.path.isdir(
+            os.path.join(
+                os.getcwd(),
+                "data/multi-goal-conditioned/multiagent-ring_small")))
+
+        # Clear anything that was generated.
+        shutil.rmtree(os.path.join(os.getcwd(), "data"))
+
+    def test_run_multi_hrl_failure_shared(self):
+        # Run the script; verify it executes without failure.
+        args = parse_train_options('', '', args=[
+            "MountainCarContinuous-v0",
+            "--shared",
+            "--initial_exploration_steps", "1",
+            "--total_steps", "500",
+            "--log_interval", "500",
+            "--alg", "woops"
+        ])
+
+        self.assertRaises(ValueError, run_multi_hrl,
+                          args=args, base_dir='data/multi-goal-conditioned')
 
         # Clear anything that was generated.
         shutil.rmtree(os.path.join(os.getcwd(), "data"))
